Remove reliance on implicit autocommit
This commit fixes the warning:

The current statement is being autocommitted using implicit autocommit,
which will be removed in SQLAlchemy 2.0. Use the .begin() method of
Engine or Connection in order to use an explicit transaction for DML
and DDL statements.
(Background on SQLAlchemy 2.0 at: https://sqlalche.me/e/b8d9)

For rationale, see section: https://docs.sqlalchemy.org/en/14/changelog/migration_20.html#library-level-but-not-driver-level-autocommit-removed-from-both-core-and-orm